Merged PR 4662: List for users
Generated new component. The list contain the status column which when is unchecked and checked it will also update the database isActive attribute. Implemented the delete functionality. Other feautures of the page will come with the enhancements list task. The css for the list should be aplied from the main style.css for all lists.

![image.png](https://dev.azure.com/smedix/9eab10cc-5d70-4301-9c8e-d24be2c89625/_apis/git/repositories/58fba364-e224-4fb6-b56f-1bb110972797/pullRequests/4662/attachments/image.png) ![image (2).png](https://dev.azure.com/smedix/9eab10cc-5d70-4301-9c8e-d24be2c89625/_apis/git/repositories/58fba364-e224-4fb6-b56f-1bb110972797/pullRequests/4662/attachments/image%20%282%29.png)

Related work items: #22404